fix [RH] : added show alert dialog

+++ lib/app/pages/detail-album-photo-page/components/detail_album_photo_page_component_one.dart
@@ -1,5 +1,6 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
+import 'package:fun_education_app_teacher/app/global-component/common_alert_dialog.dart';
 import 'package:fun_education_app_teacher/app/pages/detail-album-photo-page/detail_album_photo_controller.dart';
 import 'package:fun_education_app_teacher/common/helper/themes.dart';
 import 'package:get/get.dart';
@@ -71,8 +72,24 @@ class DetailAlbumPhotoPageComponentOne
                 Spacer(),
                 InkWell(
                   onTap: () {
-                    controller.deleteAlbumByAdmin(
-                        controller.showAllAlbumsModel.value.id!);
+                    showDialog(
+                      context: context,
+                      builder: (context) {
+                        return CommonAlertDialog(
+                          title: 'Kofirmasi',
+                          content:
+                              'Anda yakin ingin menghapus album beserta isinya?',
+                          cancelButtonText: 'Tidak',
+                          confirmButtonText: 'Iya',
+                          onConfirm: () {
+                            Get.back();
+                            controller.deleteAlbumByAdmin(
+                              controller.showAllAlbumsModel.value.id!,
+                            );
+                          },
+                        );
+                      },
+                    );
                   },
                   child: Icon(
                     Icons.delete_rounded,
